+++ packages/antwerp-ui/react-components/src/lib/molecules/card/Card.tsx
@@ -1,11 +1,20 @@
 import { CardProps } from './Card.types';
 import { Image } from '../image';
+import { classNames } from '../../../utils/dom.utils';
 import { renderHTMLLink } from '../../../utils/render.utils';
 
-export function Card({ title, subTitle, description, children, image, link, renderLinkFunction, qa }: CardProps) {
+export function Card({ title, subTitle, description, children, image, link, className, border = 'normal', highlighted = false, renderLinkFunction, qa }: CardProps) {
+  const classObject = {
+    'm-card': true,
+    'm-card--highlight': highlighted,
+    'm-card--light-border': border === 'light',
+    'm-card--no-border': border === 'none',
+    [`${className}`]: !!className
+  };
+  const classes = classNames(classObject);
   const HeaderTag = title?.tag || 'h4';
   return (
-    <div className="m-card" data-qa={qa}>
+    <div className={classes} data-qa={qa}>
       {!!image && (
         <div className="m-card__image">
           <Image {...image} />