+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 7.11.0 - 2025-10-09
+
+### Added
+
+- Added option to define border type for card
+- Added highlighted card option
+- Added option to add more classes to the card
+
+### Fixed
+
+- Fixed the alert close button that was sometimes overlapping with the title
+
 ## 7.10.3 - 2025-09-25
 
 ### Fixed